requirements.txt
# Web Framework
Flask==3.0.2
Flask-Cors==4.0.0
Flask-Migrate==4.0.5
Flask-SQLAlchemy==3.1.1
Flask-SocketIO==5.3.6
python-engineio==4.9.0
python-socketio==5.11.0
python-dotenv==1.0.1

# Database
SQLAlchemy==2.0.28
supabase==2.3.4
psycopg2-binary==2.9.9

# AI/ML
torch==2.2.1
transformers==4.38.2
datasets==2.18.0
accelerate==0.27.2
bitsandbytes==0.42.0
sentencepiece==0.2.0
protobuf==4.25.3

# System & Monitoring
psutil==5.9.8
GPUtil==1.4.0
netifaces==0.11.0
requests==2.31.0
tqdm==4.66.2

# Google Colab Integration
google-auth==2.27.0
google-auth-oauthlib==1.2.0

# Utilities
numpy==1.26.4
pandas==2.2.1
python-dateutil==2.8.2
pytz==2024.1
PyJWT==2.8.0
cryptography==42.0.5

# Development
pytest==8.0.2
black==24.2.0
flake8==7.0.0
mypy==1.8.0
isort==5.13.2
